Set item colors using a central service.
Not only is this better code, but it allow us to change how the colors
are computed.
We have introduced a simple option for selecting between basing the
colors on the Primary Historical Significance category and the CIDOC
entity class. The exact colors used for each system are then
configurable via the UI by the end user.

We take care to pass around the correct cacheability metadata so that if
colors or config are changed then the right cached items should be
expired.